fastapi/requirements.txt
groq
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
aiomysql
pymysql
cryptography